removed extra opcode file, move its contents into different relevant files
