+++ bscript/interpreter/opcodeparser_test.go
